+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/tags/base/TextTag.java
@@ -12,7 +12,8 @@
  * Lesser General Public License for more details.
  * 
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library. */
+ * License along with this library.
+ */
 package com.jpexs.decompiler.flash.tags.base;
 
 import com.jpexs.decompiler.flash.SWF;
@@ -759,7 +760,7 @@ public abstract class TextTag extends DrawableTag {
                         }
 
                         if (!"".equals(charId)) {
-                            Element charImage = exporter.addUse(mat, bounds, charId, null);
+                            Element charImage = exporter.addUse(mat, bounds, charId, null, null);
                             RGBA colorA = new RGBA(textColor);
                             charImage.setAttribute("fill", colorA.toHexRGB());
                             if (colorA.alpha != 255) {